refactor(util/ast) - util for finding closest function symbol to given location
Based on 98a21221
